#177 moved helpers.js into controllers helpers directory
